refactor: rename `isBuiltinSymbolLike.ts` -> `builtinSymbolLikes.ts`
