+++ packages/type-utils/src/builtinSymbolLikes.ts
