+++ packages/type-utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from './builtinSymbolLikes';
 export * from './containsAllTypesByName';
 export * from './getConstrainedTypeAtLocation';
 export * from './getContextualType';
@@ -6,7 +7,6 @@ export * from './getSourceFileOfNode';
 export * from './getTokenAtPosition';
 export * from './getTypeArguments';
 export * from './getTypeName';
-export * from './isBuiltinSymbolLike';
 export * from './isSymbolFromDefaultLibrary';
 export * from './isTypeReadonly';
 export * from './isUnsafeAssignment';
